Add useGetCurrentRoute hook
